linker

+++ src/main/java/fise2/hpp/ok/events/Comment.java
@@ -6,13 +6,13 @@ import fise2.hpp.ok.structs.User;
 public class Comment implements Answerable {
 
     public final String comment;
-    public int ts = 0;
+    public long ts = 0;
     public User commentator;
     public Answerable answered;
 
     public int likes = 0;
 
-    public Comment(int ts, User commentator, String comment, Answerable answered) {
+    public Comment(long ts, User commentator, String comment, Answerable answered) {
         this.ts = ts;
         this.answered = answered;
         this.comment = comment;
